Add separate profile image fields for mobile
If you use the uploader fields, it will call S3, which the phone
already did, so we’re losing the benefit of local phone processing.
Create separate pure string fields for mobile which it can set, then
return either the “web” field (if non-null), or the mobile fields in
show user.